Removes formatter
